Change Timer to use ConcurrentQueue to send Timer events
The timer thread passes messages indicating that timers have expired
to the main thread via a queue. This queue was previously guarded by
a lock which greatly hurt performance. Instead, use a ConcurrentQueue
which is implemented using atomic compare and swap operations, or on
very new platforms that support it, TSX transactions. This should
provide a reasonable performance improvement.

The queue itself is single consumer, single producer, so if a suitable
implementation of a ring was written, no locks or atomics are strictly
required. This optimization is left for the future.