Reimplement Timer system as a Hierarchical Time Wheel
Reimplement the timer system using a data structure called a
Hierarchical Time Wheel. This is a data structure for tracking
large numbers of timers with precise granularity. The change
in data structure makes tracking timers very cheap - so cheap
in fact that it is no longer necessary to offload the work to
a separate thread.

This was tested by writing an application that spawned 100k timers
of random duration and verifying that a particular chosen timer
correctly maintained its expiry time as it repeatedly expired
for several minutes. Further, the CPU usage was measured to
confirm efficiency. The test only used 4% of a single CPU.

Every attempt was made to keep the existing timer interface as is,
so very little code outside of Server/Timer.cs was modified. As of
this commit, however, the timer profiling code was removed.