Discretize time and remove Signal from main loop
Change the main loop on the main thread to attempt to run at a
fixed rate. This is done by dividing time into discrete
intervals, called ticks. In this patch, a tick is 25
milliseconds, which means the tick rate is 40Hz. Game servers
commonly run between 15Hz and 60Hz. Ultimately, everything
that happens within the same tick will be considered
simultaneous.

The purpose of the signal in the main loop was to allow the main
thread to sleep when it is idle. However, when the main thread
is not idle, checking the signal is very expensive. Instead of
optimizing for the case where the server is doing very little,
the constant tick rate optimizes for the case where the server
is very busy and performance matters.

This patch only contains the changes to the main loop. The
Core.TickCount property is still exposed and has resolution
smaller than the tick rate, since it calls
Stopwatch.GetTimestamp() each time.